Modified file structure
